+++ R/cross_fit.R
@@ -37,7 +37,7 @@
 #'   a list column named `fit`.
 #'   An alternative function can be specified with an unquoted function name or
 #'   a [purrr][purrr::map()]-style lambda function with one argument (see usage
-#'   with [broom::tidy(conf.int = TRUE)][broom::tidy()] in examples).
+#'   with [broomExtra::tidy(conf.int = TRUE)][broomExtra::tidy()] in examples).
 #'   Defaults to [tidy_glance].
 #' @param tidy_args A list of additional arguments to the `tidy` function
 #' @param errors If `"stop"`, the default, the function will stop and return an